+++ dmtool/modules/character_generator.py
